Newconf to main

+++ lib/config/Logger.ts
@@ -3,7 +3,7 @@ import * as path from 'path';
 import fs = require('fs');
 import config = require("config");
 
-if (config.get('APP.CLEAR_LOG') === true) {
+if (config.has('APP.CLEAR_LOG') && config.get('APP.CLEAR_LOG') === true) {
     try {
         fs.truncate('./logs/default.log', 0, function () { console.log('Default log clear !') });
         fs.truncate('./logs/error.log', 0, function () { console.log('Error log clear !') });
@@ -35,6 +35,7 @@ const LoggerManager = (moduleName: string) => createLogger({
     exitOnError: false,
     defaultMeta: { component: path.basename(moduleName) },
     transports: [
+        new transports.Console(),
         new transports.File({
             filename: './logs/default.log',
             level: 'info',
@@ -47,11 +48,11 @@ const LoggerManager = (moduleName: string) => createLogger({
         })
     ],
     exceptionHandlers: [
-        //new transports.Console()
+        new transports.Console(),
         new transports.File({ format: formatError, filename: './logs/error.log' })
     ],
     rejectionHandlers: [
-        //new transports.Console()
+        new transports.Console(),
         new transports.File({ format: formatError, filename: './logs/error.log' })
     ]
 });
